adicionado camada de infra e metricas

+++ internal/application/dto/product_dto.go
@@ -3,20 +3,20 @@ package dto
 import "star_store/internal/domain/entity"
 
 type InputCreateProductDto struct {
-	Title       string `json:"title"`
-	Price       int    `json:"price"`
-	ZipCode     string `json:"zip_code"`
-	Seller      string `json:"seller"`
-	ThumbnailHD string `json:"thumbnail"`
+	Title       string  `json:"title"`
+	Price       float32 `json:"price"`
+	ZipCode     string  `json:"zip_code"`
+	Seller      string  `json:"seller"`
+	ThumbnailHD string  `json:"thumbnail"`
 }
 
 type OutputCreateProductDto struct {
-	ID          string `json:"id"`
-	Title       string `json:"title"`
-	Price       int    `json:"price"`
-	ZipCode     string `json:"zip_code"`
-	Seller      string `json:"seller"`
-	ThumbnailHD string `json:"thumbnail"`
+	ID          string  `json:"id"`
+	Title       string  `json:"title"`
+	Price       float32 `json:"price"`
+	ZipCode     string  `json:"zip_code"`
+	Seller      string  `json:"seller"`
+	ThumbnailHD string  `json:"thumbnail"`
 }
 
 type OutputListProductDto struct {
