+++ internal/domain/entity/test/cart_test.go
@@ -17,7 +17,7 @@ func TestNewCart(t *testing.T) {
 	assert.Equal(t,"1236r7126",cart.ClientID)
 	assert.Equal(t,cartItems[0],cart.CartItems[0])
 	assert.Equal(t,len(cartItems),len(cart.CartItems))
-	assert.Equal(t,60.0,cart.Total)
+	assert.Equal(t,float32(60),cart.Total)
 }
 
 func TestInsertItem(t *testing.T) {
@@ -30,13 +30,13 @@ func TestInsertItem(t *testing.T) {
 	assert.NotNil(t,cart)
 	assert.Equal(t,cartItems[0].Quantity,cart.CartItems[0].Quantity)
 	assert.Equal(t,1,cart.CartItems[0].Quantity)
-	assert.Equal(t,40.0,cart.Total)
+	assert.Equal(t,float32(40),cart.Total)
 	
 	cart.InsertItem(ci1)
 	assert.Equal(t,2,cart.CartItems[0].Quantity)
-	assert.Equal(t,50.0,cart.Total)
+	assert.Equal(t,float32(50),cart.Total)
 	
 	cart.InsertItem(ci3)
 	assert.Equal(t,len(cart.CartItems),3)
-	assert.Equal(t,60.0,cart.Total)
+	assert.Equal(t,float32(60),cart.Total)
 }
